Fix File::Find installation problems
Currently, if you try to install a module that directly or indirectly
requires File::Find or other panda-required modules, panda will not
recognize that those modules were already installed in the bootstrapping
process.  As a result, unless --force is specified, the installation
will fail complaining that File::Find has already been installed.  In
addition, a directory hierarchy is created in the panda working
directory during bootstrapping under inst#, which causes other issues
when bootstrapping from the same working directory in the future.

This is because a stringy representation of a CompUnit repo is
improperly detected as a relative path, and we end up with a sort of
"split" repository between the actual install prefix and the one under
$panda_working_dir/#inst; the actual modules are installed under
the installation prefix, but the panda state is written under
$panda_working_dir/#inst.

This commit detects installation compunit repos (by checking for inst#)
and does the right thing when they are present.